Added major modifications to enhance modularity.
This is a major commit that abstracts all of the IO networking and
sercuity features into two files: src/recv_data.c, and
src/send_data.c.

The connection has been abstracted into a single connect_t structure.
This can be altered with macros to provide a different form of
security such as ssl/tls or other encryption methods or none at all.